Reorganize lend folder structure
