+++ packages/sdk/src/lend/providers/morpho/__tests__/api.test.ts
@@ -1,9 +1,9 @@
 import type { Address } from 'viem'
 import { describe, expect, it, vi } from 'vitest'
 
-import { externalTest } from '../../../utils/test.js'
-import { fetchRewards } from './api.js'
-import { fetchAndCalculateRewards } from './sdk.js'
+import { externalTest } from '../../../../utils/test.js'
+import { fetchRewards } from '../api.js'
+import { fetchAndCalculateRewards } from '../sdk.js'
 
 /**
  * External tests for API functionality
