+++ packages/sdk/src/lend/providers/morpho/__tests__/sdk.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { calculateBaseApy, calculateRewardsBreakdown } from './sdk.js'
+import { calculateBaseApy, calculateRewardsBreakdown } from '../sdk.js'
 
 describe('Vault Utilities', () => {
   describe('calculateBaseApy', () => {
